requirements.txt
langchain
langchain-google-genai
langchain-community
chromadb
requests
python-dotenv